requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
black==21.8b0
blinker==1.4
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
cryptography==3.4.8
dictpath==0.1.3
dpath==2.0.5
eight==1.0.1
flake8==3.9.2
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
Flask-Testing==0.8.1
freezegun==1.1.0
future==0.18.2
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.2.1
lazy-object-proxy==1.6.0
lxml==4.6.5
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.10.0
mypy-extensions==0.4.3
openapi-core==0.14.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
parse==1.19.0
pathspec==0.9.0
pi==0.1.2
platformdirs==2.2.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.11.0
pyflakes==2.3.1
PyJWT==2.3.0
pyOpenSSL==19.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
regex==2021.8.28
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
sentry-sdk==1.3.1
signxml==2.8.2
six==1.16.0
tma-saml==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
Werkzeug==2.0.1
xmltodict==0.12.0
zeep==4.1.0
